feat: 바뀐 디자인 및 로직 적용 (연동)
